Get metadata for a BED file or BEDset

Get metadata for a BED file or BEDset given its id. Abort if not found or id is not not 32 characters.

bb_metadata(bedbase, id, full = FALSE)

Arguments

bedbase

BEDbase() object

id

integer(1) record or object identifier

full

logical(1) (default FALSE) include full record with stats, files, and metadata

Value

list() metadata

Examples

bedbase <- BEDbase()
#> 130451 BED files available.

ex_bed <- bb_example(bedbase, "bed")
bb_metadata(bedbase, ex_bed$id)
#> $name
#> [1] "encode_1433"
#> 
#> $genome_alias
#> [1] "hg38"
#> 
#> $genome_digest
#> [1] "2230c535660fb4774114bfa966a62f823fdb6d21acf138d4"
#> 
#> $bed_compliance
#> [1] "bed6+3"
#> 
#> $data_format
#> [1] "encode_broadpeak"
#> 
#> $compliant_columns
#> [1] 6
#> 
#> $non_compliant_columns
#> [1] 3
#> 
#> $header
#> NULL
#> 
#> $id
#> [1] "dcbf6e04b953eabfe8765193a9657248"
#> 
#> $description
#> [1] "DNase-seq on human naive thymus-derived CD4-positive, alpha-beta T cell"
#> 
#> $submission_date
#> [1] "2025-04-23T22:00:32.931407Z"
#> 
#> $last_update_date
#> [1] "2025-12-23T02:30:46.287464Z"
#> 
#> $is_universe
#> [1] FALSE
#> 
#> $license_id
#> [1] "DUO:0000042"
#> 
#> $annotation
#> $annotation$species_name
#> [1] "Homo sapiens"
#> 
#> $annotation$species_id
#> [1] "9606"
#> 
#> $annotation$genotype
#> [1] ""
#> 
#> $annotation$phenotype
#> [1] ""
#> 
#> $annotation$description
#> [1] ""
#> 
#> $annotation$cell_type
#> [1] "naive thymus-derived CD4-positive, alpha-beta T cell"
#> 
#> $annotation$cell_line
#> [1] ""
#> 
#> $annotation$tissue
#> [1] ""
#> 
#> $annotation$library_source
#> [1] ""
#> 
#> $annotation$assay
#> [1] "DNase-seq"
#> 
#> $annotation$antibody
#> [1] ""
#> 
#> $annotation$target
#> [1] ""
#> 
#> $annotation$treatment
#> [1] ""
#> 
#> $annotation$global_sample_id
#> $annotation$global_sample_id[[1]]
#> [1] "encode:ENCFF823URV"
#> 
#> 
#> $annotation$global_experiment_id
#> $annotation$global_experiment_id[[1]]
#> [1] "encode"
#> 
#> 
#> $annotation$original_file_name
#> [1] "ENCFF823URV.bed.gz"
#> 
#> 
#> $processed
#> [1] TRUE
#> 
#> $stats
#> NULL
#> 
#> $plots
#> NULL
#> 
#> $files
#> NULL
#> 
#> $universe_metadata
#> NULL
#> 
#> $raw_metadata
#> NULL
#> 
#> $bedsets
#> list()
#> 

ex_bedset <- bb_example(bedbase, "bedset")
bb_metadata(bedbase, ex_bedset$id)
#> $id
#> [1] "gse125061"
#> 
#> $name
#> [1] "gse125061"
#> 
#> $md5sum
#> [1] "6dcc9a18dc9965cfc14abe7fa39d478c"
#> 
#> $submission_date
#> [1] "2025-05-23T12:16:22.281419Z"
#> 
#> $last_update_date
#> [1] "2025-05-23T12:16:22.281424Z"
#> 
#> $statistics
#> NULL
#> 
#> $plots
#> NULL
#> 
#> $description
#> [1] "Data from [GEO GSE125061](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE125061)\nThe transition from quiescent to activated states in human hematopoietic stem cells is governed by dynamic 3D genome reorganization [ChIP-Seq]"
#> 
#> $summary
#> NULL
#> 
#> $bed_ids
#> $bed_ids[[1]]
#> [1] "0692dfe199d26353af23dcbf354a6548"
#> 
#> $bed_ids[[2]]
#> [1] "36baa7aead2bad3062342fafad6c7197"
#> 
#> 
#> $author
#> [1] "Alexander, James, Murison"
#> 
#> $source
#> [1] "gse125061"
#>