An R client for BEDbase

bedbaser exposes the BEDbase API and includes convenience functions for common tasks, including importing a BED id into a GRanges object and a BEDset id into a GRangesList.

bedbaser()

Value

BEDbase object

Details

The main functions are as follows

bedbaser::bedbaser(): API constructor

bedbaser::bb_example(): Retrieve an example BED file or BEDset

bedbaser::bb_metadata(): Retrieve metadata for a BED file or BEDset

bedbaser::bb_list_beds(): List all BED files

bedbaser::bb_list_bedsets(): List all BEDsets

bedbaser::bb_beds_in_bedset(): List BED files in BEDset

bedbaser::bb_bed_text_search(): Search BED files by text

bedbaser::bb_to_granges(): Create a GRanges object from a BED id

bedbaser::bb_to_grangeslist(): Create a GrangesList from a BEDset id Set the environment variable BEDBASER_CACHE to alter the cache path for downloaded BED files.

Examples

bedbaser()
#> service: BEDbase
#> tags(); use bedbase$<tab completion>:
#> # A tibble: 32 × 3
#>    tag   operation                                                       summary
#>    <chr> <chr>                                                           <chr>  
#>  1 base  get_bedbase_db_stats_v1_stats_get                               Get su…
#>  2 base  service_info_v1_service_info_get                                GA4GH …
#>  3 bed   bed_to_bed_search_v1_bed_search_bed_post                        Search…
#>  4 bed   get_bed_classification_v1_bed__bed_id__metadata_classification… Get cl…
#>  5 bed   get_bed_embedding_v1_bed__bed_id__embedding_get                 Get em…
#>  6 bed   get_bed_files_v1_bed__bed_id__metadata_files_get                Get me…
#>  7 bed   get_bed_metadata_v1_bed__bed_id__metadata_get                   Get me…
#>  8 bed   get_bed_pephub_v1_bed__bed_id__metadata_raw_get                 Get ra…
#>  9 bed   get_bed_plots_v1_bed__bed_id__metadata_plots_get                Get pl…
#> 10 bed   get_bed_stats_v1_bed__bed_id__metadata_stats_get                Get st…
#> # ℹ 22 more rows
#> tag values:
#>   base, bed, bedset, home, objects, search, NA
#> schemas():
#>