Display API

Display API

# S4 method for class 'BEDbase'
operations(x, ..., .deprecated = FALSE)

Arguments

x

BEDbase object

...

other options

.deprecated

(default FALSE) if deprecated

Value

list() API end points

Examples

api <- BEDbase()
operations(api)
#> $lending_page__get
#> lending_page__get 
#> Lending Page 
#> 
#> Parameters:
#> 
#> $index_v1_get
#> index_v1_get 
#> API intro page 
#> Description:
#>   Display the index UI page
#> 
#> Parameters:
#> 
#> $get_example_bed_record_v1_bed_example_get
#> get_example_bed_record_v1_bed_example_get 
#> Get example BED record metadata 
#> Description:
#>   Get metadata for an example BED record.
#> 
#> Parameters:
#> 
#> $list_beds_v1_bed_list_get
#> list_beds_v1_bed_list_get 
#> Paged list of all BED records 
#> Description:
#>   Returns list of BED files in the database with optional filters.
#> 
#> Parameters:
#>   limit (integer)
#>   offset (integer)
#>   genome (string)
#>     filter by genome of the bed file. e.g. 'hg38'
#>   bed_type (string)
#>     filter by bed type. e.g. 'bed6+4'
#> 
#> $bed_to_bed_search_v1_bed_search_bed_post
#> bed_to_bed_search_v1_bed_search_bed_post 
#> Search for similar bed files 
#> 
#> Parameters:
#>   limit (integer)
#>   offset (integer)
#>   body ()
#> 
#> $text_to_bed_search_v1_bed_search_text_post
#> text_to_bed_search_v1_bed_search_text_post 
#> Search for a BedFile 
#> Description:
#>   Search for a BedFile by a text query. Example: query="cancer"
#> 
#> Parameters:
#>   query ()
#>   limit (integer)
#>   offset (integer)
#> 
#> $get_bed_embedding_v1_bed__bed_id__embedding_get
#> get_bed_embedding_v1_bed__bed_id__embedding_get 
#> Get embeddings for a single BED record 
#> Description:
#>   Returns embeddings for a single BED record.
#> 
#> Parameters:
#>   bed_id (string)
#>     BED digest
#> 
#> $get_bed_metadata_v1_bed__bed_id__metadata_get
#> get_bed_metadata_v1_bed__bed_id__metadata_get 
#> Get metadata for a single BED record 
#> Description:
#>   Example bed_id: bbad85f21962bb8d972444f7f9a3a932
#> 
#> Parameters:
#>   bed_id (string)
#>     BED digest
#>   full ()
#>     Return full record with stats, plots, files and metadata
#> 
#> $get_bed_classification_v1_bed__bed_id__metadata_classification_get
#> get_bed_classification_v1_bed__bed_id__metadata_classification_get 
#> Get classification of single BED file 
#> Description:
#>   Example bed_id: bbad85f21962bb8d972444f7f9a3a932
#> 
#> Parameters:
#>   bed_id (string)
#>     BED digest
#> 
#> $get_bed_files_v1_bed__bed_id__metadata_files_get
#> get_bed_files_v1_bed__bed_id__metadata_files_get 
#> Get metadata for a single BED record 
#> Description:
#>   Example bed_id: bbad85f21962bb8d972444f7f9a3a932
#> 
#> Parameters:
#>   bed_id (string)
#>     BED digest
#> 
#> $get_bed_plots_v1_bed__bed_id__metadata_plots_get
#> get_bed_plots_v1_bed__bed_id__metadata_plots_get 
#> Get plots for a single BED record 
#> Description:
#>   Example bed_id: bbad85f21962bb8d972444f7f9a3a932
#> 
#> Parameters:
#>   bed_id (string)
#>     BED digest
#> 
#> $get_bed_pephub_v1_bed__bed_id__metadata_raw_get
#> get_bed_pephub_v1_bed__bed_id__metadata_raw_get 
#> Get raw metadata for a single BED record 
#> Description:
#>   Returns raw metadata for a single BED record. This metadata is stored
#>   in PEPHub. And is not verified.Example bed_id:
#>   bbad85f21962bb8d972444f7f9a3a932
#> 
#> Parameters:
#>   bed_id (string)
#>     BED digest
#> 
#> $get_bed_stats_v1_bed__bed_id__metadata_stats_get
#> get_bed_stats_v1_bed__bed_id__metadata_stats_get 
#> Get stats for a single BED record 
#> Description:
#>   Example bed_id: bbad85f21962bb8d972444f7f9a3a932
#> 
#> Parameters:
#>   bed_id (string)
#>     BED digest
#> 
#> $get_regions_for_bedfile_v1_bed__bed_id__regions__chr_num__get
#> get_regions_for_bedfile_v1_bed__bed_id__regions__chr_num__get 
#> Get regions from a BED file that overlap a query region. 
#> Description:
#>   Returns the queried regions with provided ID and optional query
#>   parameters
#> 
#> Parameters:
#>   bed_id (string)
#>     BED digest
#>   chr_num (string)
#>     Chromosome number
#>   start ()
#>     query range: start coordinate
#>   end ()
#>     query range: start coordinate
#> 
#> $get_tokens_v1_bed__bed_id__tokens__universe_id__get
#> get_tokens_v1_bed__bed_id__tokens__universe_id__get 
#> Get tokenized of bed file 
#> Description:
#>   Return univers of bed file Example: bed:
#>   0dcdf8986a72a3d85805bbc9493a1302 | universe:
#>   58dee1672b7e581c8e1312bd4ca6b3c7
#> 
#> Parameters:
#>   bed_id (string)
#>   universe_id (string)
#> 
#> $get_tokens_v1_bed__bed_id__tokens__universe_id__info_get
#> get_tokens_v1_bed__bed_id__tokens__universe_id__info_get 
#> Get link to tokenized bed file 
#> Description:
#>   Return link to tokenized bed file Example: bed:
#>   0dcdf8986a72a3d85805bbc9493a1302 | universe:
#>   58dee1672b7e581c8e1312bd4ca6b3c7
#> 
#> Parameters:
#>   bed_id (string)
#>   universe_id (string)
#> 
#> $get_example_bedset_record_v1_bedset_example_get
#> get_example_bedset_record_v1_bedset_example_get 
#> Get metadata for an example BEDset record 
#> 
#> Parameters:
#> 
#> $list_bedsets_v1_bedset_list_get
#> list_bedsets_v1_bedset_list_get 
#> Paged list of all BEDset records 
#> Description:
#>   Returns a list of BEDset records in the database with optional
#>   filters and search.
#> 
#> Parameters:
#>   query (string)
#>   limit (integer)
#>   offset (integer)
#> 
#> $get_bedfiles_in_bedset_v1_bedset__bedset_id__bedfiles_get
#> get_bedfiles_in_bedset_v1_bedset__bedset_id__bedfiles_get 
#> Get Bedfiles In Bedset 
#> Description:
#>   Example bed_id: gse218680
#> 
#> Parameters:
#>   bedset_id (string)
#> 
#> $get_bedset_metadata_v1_bedset__bedset_id__metadata_get
#> get_bedset_metadata_v1_bedset__bedset_id__metadata_get 
#> Get all metadata for a single BEDset record 
#> Description:
#>   Example bed_id: gse218680
#> 
#> Parameters:
#>   bedset_id (string)
#>   full (boolean)
#> 
#> $get_bedset_metadata_v1_bedset__bedset_id__metadata_plots_get
#> get_bedset_metadata_v1_bedset__bedset_id__metadata_plots_get 
#> Get plots for single bedset record 
#> Description:
#>   Example bed_id: gse218680
#> 
#> Parameters:
#>   bedset_id (string)
#> 
#> $get_bedset_metadata_v1_bedset__bedset_id__metadata_stats_get
#> get_bedset_metadata_v1_bedset__bedset_id__metadata_stats_get 
#> Get stats for a single BEDSET record 
#> Description:
#>   Example bed_id: gse218680
#> 
#> Parameters:
#>   bedset_id (string)
#> 
#> $changelog_v1_docs_changelog_get
#> changelog_v1_docs_changelog_get 
#> Release notes 
#> 
#> Parameters:
#> 
#> $get_drs_object_metadata_v1_objects__object_id__get
#> get_drs_object_metadata_v1_objects__object_id__get 
#> Get DRS object metadata 
#> Description:
#>   Returns metadata about a DrsObject.
#> 
#> Parameters:
#>   object_id (string)
#> 
#> $get_object_bytes_url_v1_objects__object_id__access__access_id__get
#> get_object_bytes_url_v1_objects__object_id__access__access_id__get 
#> Get URL where you can retrieve files 
#> Description:
#>   Returns a URL that can be used to fetch the bytes of a DrsObject.
#> 
#> Parameters:
#>   object_id (string)
#>   access_id (string)
#> 
#> $get_object_bytes_v1_objects__object_id__access__access_id__bytes_get
#> get_object_bytes_v1_objects__object_id__access__access_id__bytes_get 
#> Download actual file 
#> Description:
#>   Returns the bytes of a DrsObject.
#> 
#> Parameters:
#>   object_id (string)
#>   access_id (string)
#> 
#> $get_object_thumbnail_v1_objects__object_id__access__access_id__thumbnail_get
#> get_object_thumbnail_v1_objects__object_id__access__access_id__thumbnail_get 
#> Download thumbnail file 
#> Description:
#>   Returns the bytes of a thumbnail of a DrsObject
#> 
#> Parameters:
#>   object_id (string)
#>   access_id (string)
#> 
#> $service_info_v1_service_info_get
#> service_info_v1_service_info_get 
#> GA4GH service info 
#> Description:
#>   Returns information about this service, such as versions, name, etc.
#> 
#> Parameters:
#> 
#> $get_bedbase_db_stats_v1_stats_get
#> get_bedbase_db_stats_v1_stats_get 
#> Get summary statistics for the DRS object store 
#> Description:
#>   Returns statistics
#> 
#> Parameters:
#>