An R client for BEDbase

bedbaser exposes the bedhost API and includes convenience functions for common tasks, such as to import a BED file by id into a GRanges object and a BEDset by its id into a GRangesList.

BEDbase(cache_path, quietly = FALSE)

getCache(x, cache_type = c("bedfiles", "bedsets"))

setCache(x, cache_path, quietly = TRUE)

Arguments

cache_path

character(1)

quietly

logical(1) (default TRUE) display messages

x

BEDbase() object

cache_type

character(1) bedfiles or bedsets

Value

BEDbase object

Details

BEDbase() creates a cache similar to that of the Geniml BBClient's cache: cache_path bedfiles a/f/afile.bed.gz bedsets a/s/aset.txt The convenience functions are as follows

bedbaser::BEDbase(): API service constructor

bedbaser::getCache(): Retrieve cache

bedbaser::setCache(): Set path to cache

bedbaser::bb_example(): Retrieve an example BED file or BEDset

bedbaser::bb_metadata(): Retrieve metadata for a BED file or BEDset

bedbaser::bb_list_beds(): List all BED files

bedbaser::bb_list_bedsets(): List all BEDsets

bedbaser::bb_beds_in_bedset(): List BED files in BEDset

bedbaser::bb_bed_text_search(): Search BED files by text

bedbaser::bb_to_granges(): Create a GRanges object from a BED id

bedbaser::bb_to_grangeslist(): Create a GRangesList from a BEDset id

bedbaser::bb_save(): Save a BED file to a path.

Examples

bedbase <- BEDbase(cache_path = tempdir())
#> 21093 BED files available.
ex_bed <- bb_example(bedbase, "bed")
bb_metadata(bedbase, ex_bed$id)
#> $name
#> [1] "LNCaP_AR_NSD2KO"
#> 
#> $genome_alias
#> [1] "hg38"
#> 
#> $genome_digest
#> [1] "2230c535660fb4774114bfa966a62f823fdb6d21acf138d4"
#> 
#> $bed_type
#> [1] "bed3+0"
#> 
#> $bed_format
#> [1] "bed"
#> 
#> $id
#> [1] "233479aab145cffe46221475d5af5fae"
#> 
#> $description
#> [1] ""
#> 
#> $submission_date
#> [1] "2024-11-18T23:02:50.801914Z"
#> 
#> $last_update_date
#> [1] "2024-11-18T23:02:50.801922Z"
#> 
#> $is_universe
#> [1] FALSE
#> 
#> $license_id
#> [1] "DUO:0000042"
#> 
#> $annotation
#> $annotation$species_name
#> [1] "Homo sapiens"
#> 
#> $annotation$species_id
#> [1] "9606"
#> 
#> $annotation$genotype
#> [1] ""
#> 
#> $annotation$phenotype
#> [1] ""
#> 
#> $annotation$description
#> [1] ""
#> 
#> $annotation$cell_type
#> [1] "metatatic castration resistant prostate cancer; AR-positive"
#> 
#> $annotation$cell_line
#> [1] "LNCaP"
#> 
#> $annotation$tissue
#> [1] "prostate"
#> 
#> $annotation$library_source
#> [1] "genomic"
#> 
#> $annotation$assay
#> [1] "ChIP-Seq"
#> 
#> $annotation$antibody
#> [1] "AR (Abcam: ab133273)"
#> 
#> $annotation$target
#> [1] ""
#> 
#> $annotation$treatment
#> [1] "Transiently with NSD2-targeting sgRNA pool"
#> 
#> $annotation$global_sample_id
#> [1] "geo:gsm7767862"
#> 
#> $annotation$global_experiment_id
#> [1] "geo:gse242734"
#> 
#> 
#> $stats
#> NULL
#> 
#> $plots
#> NULL
#> 
#> $files
#> NULL
#> 
#> $universe_metadata
#> NULL
#> 
#> $raw_metadata
#> NULL
#> 
#> $bedsets
#> list()
#>