Get metadata for a BED file or BEDset

Get metadata for a BED file or BEDset given its id. Abort if not found or id is not not 32 characters.

bb_metadata(bedbase, id, full = FALSE)

Arguments

bedbase

BEDbase() object

id

integer(1) record or object identifier

full

logical(1) (default FALSE) include full record with stats, files, and metadata

Value

list() metadata

Examples

bedbase <- BEDbase()
#> 168106 BED files available.

ex_bed <- bb_example(bedbase, "bed")
bb_metadata(bedbase, ex_bed$id)
#> $name
#> [1] "PB bulk, 100753"
#> 
#> $genome_alias
#> [1] "hg38"
#> 
#> $genome_digest
#> [1] "Ba88PY52_qeifhJrgUXyin6UITdXNsg3"
#> 
#> $bed_compliance
#> [1] "bed6+4"
#> 
#> $data_format
#> [1] "encode_narrowpeak_rs"
#> 
#> $compliant_columns
#> [1] 6
#> 
#> $non_compliant_columns
#> [1] 4
#> 
#> $header
#> NULL
#> 
#> $id
#> [1] "923202a2777c0c1a13432fae8842aba0"
#> 
#> $description
#> [1] ""
#> 
#> $submission_date
#> [1] "2026-01-17T18:32:05.098881Z"
#> 
#> $last_update_date
#> [1] "2026-02-08T17:21:01.635330Z"
#> 
#> $is_universe
#> [1] FALSE
#> 
#> $license_id
#> [1] "DUO:0000042"
#> 
#> $annotation
#> $annotation$species_name
#> [1] "Homo sapiens"
#> 
#> $annotation$species_id
#> [1] "9606"
#> 
#> $annotation$genotype
#> [1] ""
#> 
#> $annotation$phenotype
#> [1] ""
#> 
#> $annotation$description
#> [1] ""
#> 
#> $annotation$cell_type
#> [1] "bulk"
#> 
#> $annotation$cell_line
#> [1] ""
#> 
#> $annotation$tissue
#> [1] "peripheral blood"
#> 
#> $annotation$library_source
#> [1] "genomic"
#> 
#> $annotation$assay
#> [1] "ATAC-seq"
#> 
#> $annotation$antibody
#> [1] ""
#> 
#> $annotation$target
#> [1] ""
#> 
#> $annotation$treatment
#> [1] ""
#> 
#> $annotation$global_sample_id
#> $annotation$global_sample_id[[1]]
#> [1] "geo:gsm7457826"
#> 
#> 
#> $annotation$global_experiment_id
#> $annotation$global_experiment_id[[1]]
#> [1] "geo:gse234237"
#> 
#> 
#> $annotation$original_file_name
#> [1] "GSM7457826_100753_peaks.narrowPeak.gz"
#> 
#> 
#> $processed
#> [1] FALSE
#> 
#> $stats
#> NULL
#> 
#> $plots
#> NULL
#> 
#> $files
#> NULL
#> 
#> $universe_metadata
#> NULL
#> 
#> $raw_metadata
#> NULL
#> 
#> $bedsets
#> list()
#> 

ex_bedset <- bb_example(bedbase, "bedset")
bb_metadata(bedbase, ex_bedset$id)
#> $id
#> [1] "gse308450"
#> 
#> $name
#> [1] "gse308450"
#> 
#> $md5sum
#> [1] "ef23ecbdf1d2795ce414f1fd7c4bf40b"
#> 
#> $submission_date
#> [1] "2025-10-10T18:26:31.305786Z"
#> 
#> $last_update_date
#> [1] "2025-10-10T18:26:31.305789Z"
#> 
#> $statistics
#> NULL
#> 
#> $plots
#> NULL
#> 
#> $description
#> [1] "Data from [GEO GSE308450](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE308450)\nHIF-2-dependent Regulation of PTHrP and Paraneoplastic Hypercalcemia in Aggressive Clear Cell Renal Cell Carcinoma [ATAC-seq]"
#> 
#> $summary
#> NULL
#> 
#> $bed_ids
#> $bed_ids[[1]]
#> [1] "06aec17f095d6e7a92795f88938e39f7"
#> 
#> $bed_ids[[2]]
#> [1] "0cf33d94b5785adb4bad98376d9e54ed"
#> 
#> $bed_ids[[3]]
#> [1] "1192ea9caffe3d58079cebc4947a93bd"
#> 
#> $bed_ids[[4]]
#> [1] "15bf002b3f64e3dd92812ecf71598720"
#> 
#> $bed_ids[[5]]
#> [1] "3197913479d572e56504ae069b1d3506"
#> 
#> $bed_ids[[6]]
#> [1] "35b1425d8db291d044d74d169fadac74"
#> 
#> $bed_ids[[7]]
#> [1] "40295a50528c250a2770af07cb0d597a"
#> 
#> $bed_ids[[8]]
#> [1] "4439b4973a5994444a534b905e43f0b7"
#> 
#> $bed_ids[[9]]
#> [1] "447202782d8a5b91421a4b48c09bd495"
#> 
#> $bed_ids[[10]]
#> [1] "52e0c821cb493397adaa986d6beff292"
#> 
#> $bed_ids[[11]]
#> [1] "66a7de9281ba6f706bccbd062276386c"
#> 
#> $bed_ids[[12]]
#> [1] "6dcf86fbddb42da62f7fc03bd5e680d7"
#> 
#> $bed_ids[[13]]
#> [1] "71e04f8ad94edd7d4bf44a207535865d"
#> 
#> $bed_ids[[14]]
#> [1] "798186720cfef4a0032b611de668f31f"
#> 
#> $bed_ids[[15]]
#> [1] "7ab8074a09f0200ef120f30affbe8c16"
#> 
#> $bed_ids[[16]]
#> [1] "87b40f7fea7b43eea4d2e19b9fd687b4"
#> 
#> $bed_ids[[17]]
#> [1] "8ecb351d1a6b9fe8c21aa07aad3293f2"
#> 
#> $bed_ids[[18]]
#> [1] "a08fcd84a2ff158ebe537514a31c05a9"
#> 
#> $bed_ids[[19]]
#> [1] "a3ed380561a3b8d95ce33931df289e32"
#> 
#> $bed_ids[[20]]
#> [1] "ab8f75bfa95b9d3213e30a53d09a0401"
#> 
#> $bed_ids[[21]]
#> [1] "b4964fa1d0bb5de49e1e60abbfa0140b"
#> 
#> $bed_ids[[22]]
#> [1] "c88b337518a1bfd54548b48da6b24bb9"
#> 
#> $bed_ids[[23]]
#> [1] "d56e0b7d82abc7093e1ac5999d934102"
#> 
#> $bed_ids[[24]]
#> [1] "d67654d74c34a615be874a0ccc3d1e88"
#> 
#> $bed_ids[[25]]
#> [1] "deb81e0d353b17545c97d6a29afa6651"
#> 
#> $bed_ids[[26]]
#> [1] "e279ea6d52c84a8eabcb1a3a453e8c8a"
#> 
#> $bed_ids[[27]]
#> [1] "e443f4d2f27eadf7da098cf7638421eb"
#> 
#> $bed_ids[[28]]
#> [1] "e81484c5f0e1ab4e5e70274f2aabad9f"
#> 
#> $bed_ids[[29]]
#> [1] "eb9029240e7ae9da6e2fc9d06da26b0d"
#> 
#> $bed_ids[[30]]
#> [1] "ec47ec6b203f42e68a4336dc63cac68c"
#> 
#> $bed_ids[[31]]
#> [1] "f2f6492e3524404711d490f228f64762"
#> 
#> $bed_ids[[32]]
#> [1] "f54ac22e4d8290d47c5207a519e45076"
#> 
#> $bed_ids[[33]]
#> [1] "f8258f0e3a23915870c0a6477645cdae"
#> 
#> $bed_ids[[34]]
#> [1] "fbdd87b8568c1c3ce2e1398be2a4b11f"
#> 
#> $bed_ids[[35]]
#> [1] "fc3da9d53087084f12a71f1bf3205406"
#> 
#> 
#> $author
#> [1] "Bingqing, Xie"
#> 
#> $source
#> [1] "gse308450"
#>