Create a GRanges object given a BED id

Create a GRanges object given a BED id. Columns and types are generated for broad and narrow peak files. Known columns and types can be passed as a named vector through extra_cols. Otherwise, bb_to_granges() attempts to determine the column type and substitute dummy column names.

bb_to_granges(bedbase, bed_id, extra_cols = NULL, quietly = TRUE)

Arguments

bedbase

BEDbase() object

bed_id

integer(1) BED record identifier

extra_cols

character() (default NULL) extra column names to construct GRanges objects

quietly

logical(1) (default TRUE) display messages

Value

GRanges

Examples

bedbase <- BEDbase()
#> 130451 BED files available.
ex_bed <- bb_example(bedbase, "bed")
bb_to_granges(bedbase, ex_bed$id)
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://p3m.dev/cran/__linux__/noble/latest
#> GRanges object with 27834 ranges and 5 metadata columns:
#>           seqnames              ranges strand |        name     score
#>              <Rle>           <IRanges>  <Rle> | <character> <numeric>
#>       [1]     chr1       268012-268120      * |        id-1        89
#>       [2]     chr1       629871-630074      * |        id-2      1000
#>       [3]     chr1       633907-634162      * |        id-3      1000
#>       [4]     chr1       778499-778920      * |        id-4       159
#>       [5]     chr1       869744-869967      * |        id-5        53
#>       ...      ...                 ...    ... .         ...       ...
#>   [27830]     chrX 155820251-155820361      * |    id-27830        25
#>   [27831]     chrY     6533979-6534105      * |    id-27831        19
#>   [27832]     chrY   10627212-10627450      * |    id-27832       195
#>   [27833]     chrY   11668329-11668565      * |    id-27833       128
#>   [27834]     chrY   56838421-56838584      * |    id-27834        33
#>           signalValue    pValue    qValue
#>             <numeric> <numeric> <numeric>
#>       [1]          -1        -1   8.92201
#>       [2]          -1        -1 100.00000
#>       [3]          -1        -1 100.00000
#>       [4]          -1        -1  15.87760
#>       [5]          -1        -1   5.28064
#>       ...         ...       ...       ...
#>   [27830]          -1        -1   2.48767
#>   [27831]          -1        -1   1.94155
#>   [27832]          -1        -1  19.45740
#>   [27833]          -1        -1  12.83130
#>   [27834]          -1        -1   3.28479
#>   -------
#>   seqinfo: 711 sequences (1 circular) from hg38 genome