Create a GRanges object given a BED id

Create a GRanges object given a BED id. Columns and types are generated for broad and narrow peak files. Known columns and types can be passed as a named vector through extra_cols. Otherwise, bb_to_granges() attempts to determine the column type and substitute dummy column names.

bb_to_granges(
  bedbase,
  bed_id,
  file_type = "bed",
  extra_cols = NULL,
  quietly = TRUE
)

Arguments

bedbase

BEDbase() object

bed_id

integer(1) BED record identifier

file_type

character(1) bed or bigbed

extra_cols

character() (default NULL) extra column names to construct GRanges objects

quietly

logical(1) (default TRUE) display messages

Value

GRanges

Examples

bedbase <- BEDbase()
#> 22265 BED files available.
ex_bed <- bb_example(bedbase, "bed")
bb_to_granges(bedbase, ex_bed$id)
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> GRanges object with 51701 ranges and 0 metadata columns:
#>           seqnames            ranges strand
#>              <Rle>         <IRanges>  <Rle>
#>       [1]     chr1        9998-10232      *
#>       [2]     chr1     778492-778869      *
#>       [3]     chr1     815471-815747      *
#>       [4]     chr1     827327-827614      *
#>       [5]     chr1     865483-865708      *
#>       ...      ...               ...    ...
#>   [51697]     chrY 11643569-11643878      *
#>   [51698]     chrY 15948275-15948667      *
#>   [51699]     chrY 26670300-26671222      *
#>   [51700]     chrY 26671523-26671759      *
#>   [51701]     chrY 56685447-56685717      *
#>   -------
#>   seqinfo: 97 sequences from hg38 genome; no seqlengths