Create a GRanges object given a BED id

Create a GRanges object given a BED id. Columns and types are generated for broad and narrow peak files. Known columns and types can be passed as a named vector through extra_cols. Otherwise, bb_to_granges() attempts to determine the column type and substitute dummy column names.

bb_to_granges(bedbase, bed_id, extra_cols = NULL, quietly = TRUE)

Arguments

bedbase

BEDbase() object

bed_id

integer(1) BED record identifier

extra_cols

character() (default NULL) extra column names to construct GRanges objects

quietly

logical(1) (default TRUE) display messages

Value

GRanges

Examples

bedbase <- BEDbase()
#> 125948 BED files available.
ex_bed <- bb_example(bedbase, "bed")
bb_to_granges(bedbase, ex_bed$id)
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://p3m.dev/cran/__linux__/noble/latest
#> GRanges object with 76743 ranges and 6 metadata columns:
#>           seqnames            ranges strand |                   name     score
#>              <Rle>         <IRanges>  <Rle> |            <character> <numeric>
#>       [1]     chr1     181372-181622      * | Batch5_Ctrl.CBE69_S7..        29
#>       [2]     chr1     182742-182957      * | Batch5_Ctrl.CBE69_S7..        23
#>       [3]     chr1     779788-780045      * | Batch5_Ctrl.CBE69_S7..        29
#>       [4]     chr1     803090-803352      * | Batch5_Ctrl.CBE69_S7..        23
#>       [5]     chr1     812650-813214      * | Batch5_Ctrl.CBE69_S7..        23
#>       ...      ...               ...    ... .                    ...       ...
#>   [76739]     chrY 13854516-13854736      * | Batch5_Ctrl.CBE69_S7..        29
#>   [76740]     chrY 19741502-19741707      * | Batch5_Ctrl.CBE69_S7..        68
#>   [76741]     chrY 19743554-19743780      * | Batch5_Ctrl.CBE69_S7..        39
#>   [76742]     chrY 20577443-20577906      * | Batch5_Ctrl.CBE69_S7..        64
#>   [76743]     chrY 21676247-21676453      * | Batch5_Ctrl.CBE69_S7..        29
#>           signalValue    pValue    qValue      peak
#>             <numeric> <numeric> <numeric> <integer>
#>       [1]      4.2183   5.51043   2.96894       126
#>       [2]      3.7496   4.60465   2.31704       209
#>       [3]      4.2183   5.51043   2.96894        99
#>       [4]      3.7496   4.60465   2.31704       238
#>       [5]      3.7496   4.60465   2.31704       502
#>       ...         ...       ...       ...       ...
#>   [76739]     4.21830   5.51043   2.96894       105
#>   [76740]     6.13502  10.67560   6.81940       118
#>   [76741]     4.60126   6.84656   3.96129       138
#>   [76742]     6.28907  10.19680   6.46818       254
#>   [76743]     4.21830   5.51043   2.96894       113
#>   -------
#>   seqinfo: 711 sequences (1 circular) from hg38 genome